# Conflicts:
#	core/ecc.cpp

+++ core/block_rw.cpp
@@ -42,7 +42,7 @@ namespace beam
 		m_bRead = bRead;
 
 		if (bRead)
-			for (int i = 0; i < _countof(m_pS); i++)
+			for (int i = 0; i < static_cast<int>(_countof(m_pS)); i++)
 				OpenInternal(i);
 	}
 
@@ -55,7 +55,7 @@ namespace beam
 
 	void Block::BodyBase::RW::Delete()
 	{
-		for (int i = 0; i < _countof(m_pS); i++)
+		for (int i = 0; i < static_cast<int>(_countof(m_pS)); i++)
 		{
 			std::string s;
 			GetPath(s, i);
